refactor: use MVC
